requirements.txt
smspdudecoder==2.1.0
python-telegram-bot==21.1.1
python-dotenv==1.0.1
pyserial~=3.5
gpiod~=2.1.3
termcolor~=2.4.0
gpiozero~=2.0.1
lgpio~=0.2.2.0